[go-gen] Refactor main.go struct generation into own file
